DCAC-68: Confine Scala wierdness to IT test code only.
